backend/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
attrs==19.1.0
awscli==1.16.105
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
boost==0.1
botocore==1.12.137
Brlapi==0.6.4
certifi==2019.3.9
cffi==1.12.3
chainer==5.4.0
chardet==3.0.4
checkbox-support==0.22
chrome-gnome-shell==0.0.0
Click==7.0
cntk-gpu==2.7
colorama==0.3.9
command-not-found==0.3
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
dlib==19.17.0
docutils==0.14
editdistance==0.5.3
entrypoints==0.3
feedparser==5.1.3
filelock==3.0.10
Flask==1.0.3
Flask-Cors==3.0.7
future==0.17.1
gast==0.2.2
gin-config==0.1.2
google-pasta==0.1.5
grpcio==1.20.0
guacamole==0.9.2
h5py==2.9.0
html5lib==0.9999999
http-ece==1.1.0
httplib2==0.9.1
hypothesis==4.5.8
idna==2.8
imageio==2.5.0
imagesize==1.1.0
ipykernel==5.1.0
ipyparallel==6.2.3
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka-python==1.4.6
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
language-selector==0.1
lazy-object-proxy==1.3.1
leveldb==0.194
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.1
MarkupSafe==1.1.1
Mastodon.py==1.3.1
matplotlib==3.0.3
mccabe==0.6.1
menulibre==2.2.0
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==5.7.8
numpy==1.16.3
oauthlib==1.0.3
-e git+https://github.com/tensorflow/models.git@ae699073fd4c0fd57d9590ba193d0d9e8294d81b#egg=object_detection&subdirectory=research
onboard==1.2.0
opencv-python==4.1.0.25
packaging==19.0
padme==1.1.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plainbox==0.25
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==3.4.2
ptyprocess==0.6.0
pyasn1==0.1.9
pycparser==2.19
pycups==1.9.73
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
PyJWT==1.3.0
pylint==2.2.2
pyparsing==2.4.0
pyrsistent==0.14.11
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.0
python-debian==0.1.27
python-gflags==3.1.2
python-systemd==231
pytz==2019.1
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
reportlab==3.3.0
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.3
sqlparse==0.3.0
system-service==0.3
tb-nightly==1.14.0a20190301
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-serving-api==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.1.0a0+309c15e
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.1
typing==3.6.6
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24.2
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.11.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
